Adding the first platformer DEMO!
This commit adds a ton of objects that I will need to weed out here in a
bit. It is a simple platform with a character and two "choice" objects
(a crate and a drum).

You can control the character in the Demo0001 scene and interact with
the objects by pushing them around. The code that makes them able to
influence the game state still needs to be added.

Will need to confer with izach about the how on that.