Sexing up XmlUtilities to allow for generic replace/transform methods.
It think this is a good solution to our longterm problem of parsing more structured data from the xml.

This refactors the <regex, delegate> dictionary to be a <regex, string> dictionary. The strings are the names of the
methods one wishes to invoke when a data element matches regex. When a data string is found
to match regex this method uses reflection to lookup the method in the XmlTransforms
class and then invoke that method.

The benefit being that we don't have to be tied into a specific delegate type for replacers.

The disadvantage being that reflection style code cannot be caught by the compiler and may lead to Null pointer exceptions.


Unfortunately C# doesn't have good ol' void * for functions so this was created in an attempt to make sure that the
XML parsing core can be situated in one logical location rather than spaced around in specific code files.