SolutionManager almost done.
Last part is mapping the guesses onto the solution list in such a way you can tell if there is a combined solution.

Guesses.select<>(g => solutions.any(g => isBull(g)).All()?

For each guess see if it is a bull in any solution, if every guess is in a solution then it has been solved!