Prefab update to round off previous commits.
A change that you can't see since it is tucked inside the binary of the prefab/scene is that I adjusted the ortho projection of the BC camera a bit.

It now should be more "zoomed in" and the pieces will be larger because there was a metric shit ton of unused screen real estate.